✨ 그룹 참여 동작을 그룹 참여 페이지에서 처리하도록 코드 이동
